requirements.txt
aiofiles==0.7.0
aioredis==2.0.0
aiosmtplib==1.1.6
alembic==1.7.5
amqp==5.0.9
anyio==3.3.4
asgiref==3.4.1
astroid==2.11.7
async-timeout==4.0.1
attrs==21.4.0
autopep8==1.6.0
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
blinker==1.4
boto3==1.20.26
botocore==1.23.26
cached-property==1.5.2
cachetools==5.2.0
celery==5.2.3
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
dill==0.3.5.1
distlib==0.3.4
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.3
fakeredis==1.7.0
fastapi==0.70.0
fastapi-mail==1.0.2
ffmpeg-python==0.2.0
filelock==3.4.0
flake8==4.0.1
flower==1.0.0
future==0.18.2
google-api-core==2.8.2
google-auth==2.9.1
google-cloud==0.34.0
google-cloud-core==2.3.1
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
humanize==3.14.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
iniconfig==1.1.1
install==1.3.5
isort==5.10.1
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.2.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.3
passlib==1.7.4
pep8==1.7.1
platformdirs==2.4.0
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==4.21.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
pylint==2.14.4
PyMySQL==1.0.2
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
redis==4.0.2
requests==2.27.1
rfc3986==1.5.0
rsa==4.8
s3transfer==0.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.27
starlette==0.16.0
stripe==2.64.0
termcolor==1.1.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
tornado==6.1
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
wrapt==1.13.3
zipp==3.6.0
